+++ backend/src/features/auth/validators/index.ts
@@ -1,0 +1,2 @@
+export * from "./email.validators.js";
+export * from "./password.validators.js";
